cardslide activity can be seen again, cards need to be dealt by server + presets

+++ appagainsthumanity/src/at/tugraz/iicm/ma/appagainsthumanity/db/SetterProxy.java
@@ -69,7 +69,7 @@ public class SetterProxy {
 		return insertIgnoreOverwrite(DBContract.User.TABLE_NAME, DBContract.User._ID, values);
 	}
 	
-	public long addUser(int id, String username) {
+	public long addUser(long id, String username) {
 		ContentValues values = new ContentValues();
 		values.put(DBContract.User._ID, id);
 		values.put(DBContract.User.COLUMN_NAME_USERNAME, username);
